pyproject.toml
[tool.poetry]
name = "undou-back"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-2.0.0-cp310-none-macosx_10_9_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cpu/torchvision-0.15.1-cp310-cp310-macosx_10_9_x86_64.whl"}
uvicorn = "^0.22.0"
fastapi = "^0.95.2"
numpy = "^1.24.3"
Pillow = "^9.5.0"
matplotlib = "^3.7.1"
jax = "^0.4.11"
jaxlib = "^0.4.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
